requirements.txt
streamlit
scikit-learn
pandas
matplotlib
seaborn
xgboost
gspread
google-auth
